requirements.txt
streamlit
huggingface_hub
pandas
numpy
openpyxl
scikit-learn
sentence-transformers
scipy
openai
python-dotenv
transformers